requirements.txt
openai
typing_extensions
pyyaml